requirements.txt
#ultralytics
ultralytics-opencv-headless
opencv-python
huggingface_hub
dotenv
fastapi
uvicorn
python-multipart
